prevent UI block after copy&paste
when big text parts are inserted per copy&paste, the UI could block
because of MarkdownHighlighter::highlightBlock() call "storm".

this change prevents the blocking by checking if the document actually
changed since the last call. But now we have to reset this previousText
variable when we load a file or use the "new" action.